I messed this up
I want to have the image on one side of the home page/index.html page and I want the text to be on the other side with the three social links at the bottom evenly spaced. My CSS isn't linking to any class names. I'm not sure how to fix it, and the about page seems to look even worse.